Undo on the duplication work
