--- apps/front-end/src/pages/admin/PoAdmin/Common.js
@@ -1,22 +1,0 @@
-export const CustomStyles = {
-  rows: {
-    style: {
-      minHeight: "72px",
-    },
-  },
-  headCells: {
-    style: {
-      background: "#E0E0E0",
-      color: "#616161",
-      size: "16px",
-      justifyContent: "center", // override the alignment of columns
-    },
-  },
-  cells: {
-    style: {
-      color: "#616161",
-      size: "19px",
-      justifyContent: "center", // override the alignment of columns
-    },
-  },
-};
