+++ apps/front-end/src/pages/admin/PoAdmin/IP/List.js
@@ -13,9 +13,31 @@ import { debounce } from "lodash";
 import DataTable from "react-data-table-component";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom";
-import { CustomStyles } from "../Common";
 
-const columns = (t) => [
+export const CustomStyles = {
+  rows: {
+    style: {
+      minHeight: "72px",
+    },
+  },
+  headCells: {
+    style: {
+      background: "#E0E0E0",
+      color: "#616161",
+      size: "16px",
+      justifyContent: "center", // override the alignment of columns
+    },
+  },
+  cells: {
+    style: {
+      color: "#616161",
+      size: "19px",
+      justifyContent: "center", // override the alignment of columns
+    },
+  },
+};
+
+const columns = (t, navigate) => [
   {
     name: t("IP_ID"),
     selector: (row) => row?.id,
